Implement preact/idom strategies to rerender
