+++ package.json
@@ -71,6 +71,7 @@
   "dependencies": {
     "delegate-it": "^1.1.0",
     "delegated-events": "^1.1.1",
+    "domdiff": "^2.1.0",
     "fast-deep-equal": "^2.0.1",
     "fast-on-load": "^1.2.0",
     "htm": "^2.2.1",
